requirements.txt
tensorflow-macos
tensorflow-metal
mediapipe
opencv-python
numpy
